requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
filelock==3.18.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hf-xet==1.0.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kafka-python==2.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.36.0
networkx==3.4.2
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
portalocker==2.10.1
praw==7.8.1
prawcore==2.4.0
protobuf==5.29.4
py4j==0.10.9.7
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pymongo==4.12.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.44.1
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
update-checker==0.18.0
urllib3==2.4.0
watchdog==6.0.0
websocket-client==1.8.0
